MAINT: Move `InputRejected` to `IPython.core.error`.
