+++ IPython/core/error.py
@@ -51,3 +51,10 @@ class StdinNotImplementedError(IPythonCoreError, NotImplementedError):
     For use in IPython kernels, where only some frontends may support
     stdin requests.
     """
+
+class InputRejected(Exception):
+    """Input rejected by ast transformer.
+
+    Raise this in your NodeTransformer to indicate that InteractiveShell should
+    not execute the supplied input.
+    """
